No service/probes on agent with ZBX_PASSIVE_ALLOW false
When you're actively trying to disable the passive agent, liveness
probes should not be conducted, and there's no point in creating a
service.

The default values file sets it to true and by omission (in the case
of overriding the env) it will still default to true in the docker
image.

So the only way to affect this is setting ZBX_PASSIVE_ALLOW=false,
any other variation or omission will result in true, hence the check
becomes the somewhat convoluted "not has ZBX_PASSIVE_ALLOW false"